x86/intel_rdt.c: Disable preempt for intel_rdt_sched_in() in move_myself()
intel_rdt_sched_in() calls this_cpu_ptr() read and write pqr_state and
update PQR_ASSOC on current cpu. If execution of the function is preempted
and switched to another CPU, the results are wrong. If
CONFIG_DEBUG_PREEMPT is turned on, the issue is reported as
"BUG: smp_processor_id() running in preemptible code." when moving a task
to a rdtgroup in move_myself().

Disable preempt before intel_rdt_sched_in() and eanble preempt after
the function in move_myself() to solve the issue.

Signed-off-by: Fenghua Yu <fenghua.yu@intel.com>